Adapt the fingerprint to use the device name instead of the product
The flavours are defined as different products from each other,
they therefore all have unique product names. For legacy reasons,
the product mentioned in the fingerprint has to be 'FP2' - which is
also the device name. This commit adapts the fingerprint to use the
device name instead of the product name and keep a nicer separation
for the flavours.

FP-Id: FP2N-101
Change-Id: If307a1478b6a4e170a55382d928aa05ebd624c80